+++ README.md
@@ -41,7 +41,7 @@
 ### TO-DO
 - [x] Enum 출력값 변경
   - [x] 승패 출력값 변경
-  - [x] domain.constant.Suit 출력값 변경
+  - [x] domain.card.Suit 출력값 변경
 - [x] 딜러 승패 판정 오류 수정
 - [x] 버스트인데 hit/stand 질문
 - [x] Ace, Jack, Queen, King 이름이 값(1,10)으로 출력되는 문제
