+++ src/main/java/domain/Hand.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.card.Card;
+
 import java.util.ArrayList;
 import java.util.List;
