+++ src/test/java/domain/participant/PlayersTest.java
@@ -1,13 +1,13 @@
 package domain.participant;
 
-import domain.Card;
+import domain.card.Card;
 import domain.dto.GameScoreResultDto;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static domain.constant.Rank.*;
-import static domain.constant.Suit.*;
+import static domain.card.Rank.*;
+import static domain.card.Suit.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
 import static org.junit.jupiter.api.Assertions.assertAll;
