+++ src/test/java/domain/result/ProfitTest.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.result;
 
-import domain.constant.Result;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
